+++ src/main/java/wooteco/subway/admin/controller/PageController.java
@@ -8,9 +8,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class PageController {
 
-    public PageController() {
-    }
-
     @GetMapping(value = "/", produces = MediaType.TEXT_HTML_VALUE)
     public String index() {
         return "service/index";
